Revert "Parser (#45)"
This reverts commit 2ef5bf341a69cf19feb413c1b9ed2dfddfd986e2.
