+++ Makefile
@@ -45,7 +45,6 @@ LEXSRCS += free.c
 LEXSRCS += heredoc.c
 LEXSRCS += inhib.c
 LEXSRCS += lexer.c
-LEXSRCS += misc.c
 LEXSRCS += newline.c
 LEXSRCS += quote.c
 LEXSRCS += redir.c
@@ -57,7 +56,6 @@ LEXSRCS += state/control.c
 LEXSRCS += state/exp.c
 LEXSRCS += state/flag.c
 LEXSRCS += state/hdbody.c
-LEXSRCS += state/init.c
 LEXSRCS += state/ionumber.c
 LEXSRCS += state/redir.c
 LEXSRCS += state/start.c
@@ -73,7 +71,6 @@ PARSRCS += cmd.c
 PARSRCS += free.c
 PARSRCS += init.c
 PARSRCS += lst_to_tab.c
-PARSRCS += misc.c
 PARSRCS += newline.c
 PARSRCS += parser.c
 PARSRCS += redir.c
@@ -86,16 +83,9 @@ PARSRCS += state/cmd_args.c
 PARSRCS += state/cmd_start.c
 PARSRCS += state/delim.c
 PARSRCS += state/error.c
-PARSRCS += state/init.c
 PARSRCS += state/io_number.c
 PARSRCS += state/redir.c
 PARSRCS += state/table_start.c
-
-## ANALYZER ##
-
-ANASRCS += jobs_str/core.c
-ANASRCS += jobs_str/redir.c
-ANASRCS += heredoc.c
 
 ## EVAL ##
 
@@ -119,16 +109,6 @@ EVALSRCS += tools_redir.c
 
 BTSRCS += exit.c
 
-## INCLUDES ##
-
-INCLUDES += analyzer.h
-INCLUDES += exec.h
-INCLUDES += lexer.h
-INCLUDES += line_edition.h
-INCLUDES += parser.h
-INCLUDES += sh.h
-INCLUDES += struct.h
-INCLUDES += var.h
 
 ## INCLUDES ##
 
@@ -151,7 +131,6 @@ SRC += init_cfg.c
 SRC += $(addprefix line_edition/,$(LESRCS))
 SRC += $(addprefix lexer/,$(LEXSRCS))
 SRC += $(addprefix parser/,$(PARSRCS))
-SRC += $(addprefix analyzer/,$(ANASRCS))
 SRC += $(addprefix evaluator/,$(EVALSRCS))
 SRC += $(addprefix builtins/,$(BTSRCS))
 SRC += $(addprefix debug/,$(DBSRCS))
@@ -162,8 +141,6 @@ OPATHS += $(OPATH)lexer
 OPATHS += $(OPATH)lexer/state
 OPATHS += $(OPATH)parser
 OPATHS += $(OPATH)parser/state
-OPATHS += $(OPATH)analyzer
-OPATHS += $(OPATH)analyzer/jobs_str
 OPATHS += $(OPATH)evaluator
 OPATHS += $(OPATH)builtins
 OPATHS += $(OPATH)debug
@@ -185,7 +162,7 @@ LIPATH = libft/includes/
 LIB = $(LPATH)libft.a
 LIBDB = $(LPATH)libft_db.a
 
-WFLAGS = -g -Wall -Werror -Wextra
+WFLAGS = -Wall -Werror -Wextra
 IFLAGS = -I $(IPATH) -I $(LIPATH)
 CFLAGS = $(WFLAGS) $(IFLAGS)
 DBFLAGS = -fsanitize=address
